+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project attempts to adhere to [Semantic Versioning](https://semver.org/
 
 ## [Unreleased]
 
+### ❗ Breaking Changes
+
+- Removed deprecated `PrimaryDirectoryPath`, `StorageVolumesDirectoryPaths`, and `MusicDirectoryPath`.
+  - It seems like `getTypedExportedConstants()`, which creates these constants caused crashes on some devices according to the Google Play Console.
+
 ## [0.9.0] - 2025-05-24
 
 ### ⚡ Changes